app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.34
alembic==1.13.3
PyMySQL==1.1.1
cryptography==42.0.5
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
marshmallow==3.21.2
marshmallow-sqlalchemy==0.29.0
flask-cors==4.0.0
flasgger==0.9.7b2
pytz==2024.1
passlib==1.7.4
Werkzeug==3.0.3
Faker==26.0.0
pytest==8.3.2
pytest-flask==1.3.0
requests==2.32.3
